feat: created new component for the date button to pass on needed props
